refactor: 유저 정보 수정 리팩토링

See merge request s11-webmobile1-sub2/S11P12A804!53

+++ backend/board/src/main/java/com/ssafy/ozz/board/domain/Tag.java
@@ -20,10 +20,10 @@ public class Tag {
     private Board board;
 
     @Column(name = "x_position")
-    private Float xPosition;
+    private double xPosition;
 
     @Column(name = "y_position")
-    private Float yPosition;
+    private double yPosition;
 
     @Column(name = "clothes_id")
     private Long clothesId;
